+++ CHANGELOG.md
@@ -33,6 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new command `variable-conversion.showConvertCaseOrderDialog` to display current format order configuration. (新增 `variable-conversion.showConvertCaseOrderDialog` 命令，用于显示当前配置的格式顺序信息)
 - Add clickable link in formatOrder configuration description to trigger info display. (在 formatOrder 配置项描述中添加了点击触发命令的超链接)
 - Add shortcut key `Ctrl + Alt + \` to trigger format order info display dialog. (新增快捷键 `Ctrl + Alt + \` 用于触发显示格式顺序弹窗)
+- Support displaying duplicate and invalid items in format order information dialog. (支持在格式顺序信息弹窗中显示配置的重复项和无效项)
+- Add buttons in format order dialog for quick navigation to corresponding settings. (在格式顺序弹窗中添加按钮，支持快捷跳转至设置页对应配置项)
 
 ### Improved